# cmd/api/docs/swagger.yaml
basePath: /
definitions:
...
  dto.CreateAddressRequest:
    properties:
      active:
        example: true
        type: boolean
      address:
        example: Jl. Sudirman No. 123, Jakarta Selatan
        type: string
      customer_id:
        type: integer
      isMain:
        example: true
        type: boolean
      name:
        example: Head Office
        type: string
    required:
    - address
    - customer_id
    - name
    type: object
  dto.CreateContactRequest:
    properties:
      active:
        example: true
        type: boolean
      birthdate:
        example: "1985-03-15"
        type: string
      customer_id:
        type: integer
      email:
        example: budi@digiinno.com
        type: string
      isMain:
        example: true
        type: boolean
      jobPosition:
        example: CEO
        type: string
      mobile:
        example: 0812-3456-7890
        type: string
      name:
        example: Budi Santoso
        type: string
      phone:
        example: 021-5551234
        type: string
    required:
    - customer_id
    - name
    type: object
  dto.CreateCustomerRequest:
    properties:
      accountManagerId:
        description: Ubah dari account_manager_id ke accountManagerId
        type: string
      addresses:
        items:
          $ref: '#/definitions/dto.CreateAddressRequest'
        type: array
      brandName:
        description: Ubah dari brand_name ke brandName
        type: string
      code:
        type: string
      contacts:
        items:
          $ref: '#/definitions/dto.CreateContactRequest'
        type: array
      groups:
        items:
          $ref: '#/definitions/dto.CreateGroupsRequest'
        type: array
      logo:
        type: string
      name:
        type: string
      others:
        items:
          $ref: '#/definitions/dto.CreateOtherRequest'
        type: array
      socials:
        items:
          $ref: '#/definitions/dto.CreateSocialRequest'
        type: array
      status_name:
        type: string
      structures:
        items:
          $ref: '#/definitions/dto.CreateStructureRequest'
        type: array
    required:
    - name
    type: object
  dto.CreateGroupsRequest:
    properties:
      industryActive:
        example: true
        type: boolean
      industryId:
        description: Perbaiki nama field
        example: "1"
        type: string
      parentGroupActive:
        example: true
        type: boolean
      parentGroupId:
        description: Perbaiki nama field
        example: "2"
        type: string
    type: object
  dto.CreateOtherRequest:
    properties:
      active:
        example: true
        type: boolean
      customer_id:
        type: integer
      key:
        example: company_size
        type: string
      value:
        example: 50-100 employees
        type: string
    required:
    - customer_id
    - key
    type: object
  dto.CreateSocialRequest:
    properties:
      active:
        example: true
        type: boolean
      handle:
        example: '@digiinno_id'
        type: string
      name:
        example: Instagram
        type: string
      platform:
        example: Instagram
        type: string
    required:
    - handle
    - name
    - platform
    type: object
...
  dto.CreateStructureRequest:
    properties:
      active:
        example: true
        type: boolean
      address:
        example: Jakarta
        type: string
      customer_id:
        type: integer
      level:
        example: 1
        type: integer
      name:
        example: Board of Directors
        type: string
      parentKey:
        example: "null"
        type: string
      tempKey:
        example: "1"
        type: string
    required:
    - customer_id
    - level
    - name
    type: object
  dto.Customer:
    properties:
      average_cost:
        example: 50000000
        type: number
      brand_name:
        example: TechMaju
        type: string
      category:
        example: Technology
        type: string
      code:
        example: TM001
        type: string
      description:
        example: Perusahaan teknologi informasi
        type: string
      email:
        example: info@teknologimaju.com
        type: string
      id:
        example: 1
        type: integer
      name:
        example: PT Teknologi Maju
        type: string
      phone:
        example: 021-12345678
        type: string
      rating:
        example: 4.5
        type: number
      status:
        example: Active
        type: string
      website:
        example: https://teknologimaju.com
        type: string
    type: object
...
  dto.CustomersResponse:
    properties:
      customers:
        items:
          $ref: '#/definitions/dto.Customer'
        type: array
      stats:
        $ref: '#/definitions/dto.Stats'
    type: object
  dto.ErrorResponse:
    properties:
      error:
        example: Invalid request
        type: string
    type: object
...
  dto.Stats:
    properties:
      avg_cost:
        example: 45000000
        type: number
      blocked_customers:
        example: 5
        type: integer
      new_customers:
        example: 10
        type: integer
      total_customers:
        example: 100
        type: integer
    type: object
...
  entity.Activity:
    properties:
      activity_checkins:
        items:
          $ref: '#/definitions/entity.ActivityCheckin'
        type: array
      agenda:
        type: string
      attendees:
        items:
          $ref: '#/definitions/entity.User'
        type: array
      created_at:
        type: string
      created_by:
        type: integer
      creator:
        $ref: '#/definitions/entity.User'
      customer_id:
        type: integer
      end_time:
        type: string
      id:
        type: integer
      location_name:
        type: string
      start_time:
        type: string
      status:
        type: string
      title:
        type: string
      type:
        type: string
      updated_at:
        type: string
    type: object
  entity.ActivityCheckin:
    properties:
      activity_id:
        type: integer
      checked_in_at:
        type: string
      created_at:
        type: string
      id:
        type: integer
      updated_at:
        type: string
      user:
        $ref: '#/definitions/entity.User'
      user_id:
        type: integer
    type: object
  entity.Address:
    properties:
      active:
        type: boolean
      address:
        type: string
      city:
        type: string
      country:
        type: string
      created_at:
        type: string
      customer_id:
        type: integer
      id:
        type: integer
      main:
        type: boolean
      name:
        description: SupplierID *uint          `json:"supplier_id"` // HAPUS field
          ini
        type: string
      postal_code:
        type: string
      state:
        type: string
      street:
        type: string
      updated_at:
        type: string
    type: object
  entity.Contact:
    properties:
      active:
        type: boolean
      birthdate:
        type: string
      created_at:
        type: string
      customer_id:
        type: integer
      department:
        type: string
      email:
        type: string
      id:
        type: integer
      job_position:
        type: string
      main:
        type: boolean
      mobile:
        type: string
      name:
        type: string
      phone:
        type: string
      position:
        type: string
      updated_at:
        type: string
    type: object
  entity.Customer:
    properties:
      account_manager_id:
        type: string
      activities:
        items:
          $ref: '#/definitions/entity.Activity'
        type: array
      addresses:
        description: Relations
        items:
          $ref: '#/definitions/entity.Address'
        type: array
      average_cost:
        type: number
      brand_name:
        type: string
      category:
        type: string
      code:
        type: string
      contacts:
        items:
          $ref: '#/definitions/entity.Contact'
        type: array
      created_at:
        type: string
      description:
        type: string
      email:
        type: string
      groups:
        items:
          $ref: '#/definitions/entity.Group'
        type: array
      id:
        type: integer
      logo:
        type: string
      name:
        type: string
      others:
        items:
          $ref: '#/definitions/entity.Other'
        type: array
      phone:
        type: string
      rating:
        type: number
      sosmeds:
        items:
          $ref: '#/definitions/entity.Sosmed'
        type: array
      status:
        description: Status internal
        type: string
      structures:
        items:
          $ref: '#/definitions/entity.Structure'
        type: array
      updated_at:
        type: string
      website:
        type: string
    type: object
  entity.Group:
    properties:
      active:
        type: boolean
      created_at:
        type: string
      customers:
        description: Relations
        items:
          $ref: '#/definitions/entity.Customer'
        type: array
      id:
        type: integer
      name_group:
        type: string
      updated_at:
        type: string
      value:
        type: string
    type: object
  entity.Other:
    properties:
      active:
        type: boolean
      created_at:
        type: string
      customer_id:
        type: integer
      id:
        type: integer
      key:
        type: string
      updated_at:
        type: string
      value:
        type: string
    type: object
  entity.Role:
    properties:
      created_at:
        type: string
      id:
        type: integer
      role_name:
        type: string
      updated_at:
        type: string
      users:
        description: Relations
        items:
          $ref: '#/definitions/entity.User'
        type: array
    type: object
  entity.Sosmed:
    properties:
      active:
        type: boolean
      created_at:
        type: string
      customer_id:
        type: integer
      followers:
        type: integer
      handle:
        type: string
      id:
        type: integer
      name:
        type: string
      platform:
        type: string
      updated_at:
        type: string
      url:
        type: string
      username:
        type: string
    type: object
  entity.Structure:
    properties:
      active:
        type: boolean
      address:
        type: string
      children:
        items:
          $ref: '#/definitions/entity.Structure'
        type: array
      created_at:
        type: string
      customer_id:
        type: integer
      id:
        type: integer
      level:
        type: integer
      name:
        type: string
      parent:
        $ref: '#/definitions/entity.Structure'
      parent_id:
        type: integer
      position:
        type: integer
      updated_at:
        type: string
    type: object
  entity.User:
    properties:
      activities:
        items:
          $ref: '#/definitions/entity.Activity'
        type: array
      activity_checkins:
        items:
          $ref: '#/definitions/entity.ActivityCheckin'
        type: array
      attending_activities:
        items:
          $ref: '#/definitions/entity.Activity'
        type: array
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      role:
        allOf:
        - $ref: '#/definitions/entity.Role'
        description: |-
          Update struct User, tambahkan di bagian Relations:
          Relations
      role_id:
        description: Default to regular user role
        type: integer
      updated_at:
        type: string
      username:
        type: string
    type: object
...
host: localhost:9000
info:
  contact: {}
  description: API untuk manajemen customer menggunakan Golang, Gin, dan PostgreSQL
  title: Customer Management API
  version: "1.0"
paths:
...
  /api/customers:
    get:
      consumes:
      - application/json
      description: Get list of all customers with optional status filter
      parameters:
      - description: Filter by status
        enum:
        - Active
        - Inactive
        - Blocked
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CustomersResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all customers
      tags:
      - Customers
    post:
      consumes:
      - application/json
      description: Create a new customer record with all related data including addresses,
        social media, contacts, structures, groups, and other attributes
      parameters:
      - description: Customer data
        in: body
        name: customer
        required: true
        schema:
          $ref: '#/definitions/dto.CreateCustomerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entity.Customer'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create new customer
      tags:
      - Customers
  /api/customers/{id}:
    get:
      consumes:
      - application/json
      description: Get a specific customer by ID
      parameters:
      - description: Customer ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Customer'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get customer by ID
      tags:
      - Customers
...
securityDefinitions:
  BearerAuth:
    description: Type \"Bearer\" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"